+++ precompiles/distribution/integration_test.go
@@ -20,6 +20,7 @@ import (
 	evmosutil "github.com/evmos/evmos/v18/testutil"
 	testutiltx "github.com/evmos/evmos/v18/testutil/tx"
 	"github.com/evmos/evmos/v18/utils"
+
 	//nolint:revive // dot imports are fine for Ginkgo
 	. "github.com/onsi/ginkgo/v2"
 	//nolint:revive // dot imports are fine for Ginkgo
@@ -173,6 +174,52 @@ var _ = Describe("Calling distribution precompile from EOA", func() {
 			expFinal := initialBalance.Amount.Int64() + expRewardAmt.Int64() - fees
 			Expect(finalBalance.Amount.Equal(math.NewInt(expFinal))).To(BeTrue(), "expected final balance to be equal to initial balance + rewards - fees")
 		})
+
+		It("should withdraw delegation rewards to a smart contract", func() {
+			// deploy a smart contract to use as withdrawer
+			distributionCallerContract, err := contracts.LoadDistributionCallerContract()
+			Expect(err).To(BeNil(), "error while loading the smart contract: %v", err)
+
+			contractAddr, err := s.DeployContract(distributionCallerContract)
+			Expect(err).To(BeNil(), "error while deploying the smart contract: %v", err)
+
+			initialWithdrawerBalance := s.app.BankKeeper.GetBalance(s.ctx, contractAddr.Bytes(), s.bondDenom)
+			Expect(initialWithdrawerBalance.Amount).To(Equal(sdk.ZeroInt()))
+
+			// set contract address as withdrawer address
+			err = s.app.DistrKeeper.SetWithdrawAddr(s.ctx, s.address.Bytes(), contractAddr.Bytes())
+			Expect(err).To(BeNil())
+
+			// get tx sender initial balance
+			initialBalance := s.app.BankKeeper.GetBalance(s.ctx, s.address.Bytes(), s.bondDenom)
+
+			withdrawRewardsArgs := defaultWithdrawRewardsArgs.
+				WithArgs(s.address, s.validators[0].OperatorAddress).
+				WithGasPrice(gasPrice)
+
+			withdrawalCheck := passCheck.
+				WithExpEvents(distribution.EventTypeWithdrawDelegatorRewards)
+
+			res, ethRes, err := contracts.CallContractAndCheckLogs(s.ctx, s.app, withdrawRewardsArgs, withdrawalCheck)
+			Expect(err).To(BeNil(), "error while calling the precompile")
+
+			var rewards []cmn.Coin
+			err = s.precompile.UnpackIntoInterface(&rewards, distribution.WithdrawDelegatorRewardsMethod, ethRes.Ret)
+			Expect(err).To(BeNil())
+			Expect(len(rewards)).To(Equal(1))
+			Expect(rewards[0].Denom).To(Equal(s.bondDenom))
+			Expect(rewards[0].Amount).To(Equal(expRewardAmt))
+
+			// check tx sender balance is reduced by fees paid
+			finalBalance := s.app.BankKeeper.GetBalance(s.ctx, s.address.Bytes(), s.bondDenom)
+			fees := sdk.NewIntFromBigInt(gasPrice).MulRaw(res.GasUsed)
+			expFinal := initialBalance.Amount.Sub(fees)
+			Expect(finalBalance.Amount).To(Equal(expFinal), "expected final balance to be equal to initial balance - fees")
+
+			// check that the rewards were added to the withdrawer balance
+			finalWithdrawerBalance := s.app.BankKeeper.GetBalance(s.ctx, contractAddr.Bytes(), s.bondDenom)
+			Expect(finalWithdrawerBalance.Amount.BigInt()).To(Equal(expRewardAmt))
+		})
 	})
 
 	Describe("Validator Commission: Execute WithdrawValidatorCommission tx", func() {
@@ -257,6 +304,52 @@ var _ = Describe("Calling distribution precompile from EOA", func() {
 			fees := gasPrice.Int64() * res.GasUsed
 			expFinal := initialBalance.Amount.Int64() + expCommAmt.Int64() - fees
 			Expect(finalBalance.Amount.Equal(math.NewInt(expFinal))).To(BeTrue(), "expected final balance to be equal to the final balance after withdrawing commission")
+		})
+
+		It("should withdraw validator commission to a smart contract", func() {
+			// deploy a smart contract to use as withdrawer
+			distributionCallerContract, err := contracts.LoadDistributionCallerContract()
+			Expect(err).To(BeNil(), "error while loading the smart contract: %v", err)
+
+			contractAddr, err := s.DeployContract(distributionCallerContract)
+			Expect(err).To(BeNil(), "error while deploying the smart contract: %v", err)
+
+			initialWithdrawerBalance := s.app.BankKeeper.GetBalance(s.ctx, contractAddr.Bytes(), s.bondDenom)
+			Expect(initialWithdrawerBalance.Amount).To(Equal(sdk.ZeroInt()))
+
+			// set contract address as withdrawer address
+			err = s.app.DistrKeeper.SetWithdrawAddr(s.ctx, s.address.Bytes(), contractAddr.Bytes())
+			Expect(err).To(BeNil())
+
+			// initial balance should be the initial amount minus the staked amount used to create the validator, minus fees paid for deploying contract
+			initialBalance := s.app.BankKeeper.GetBalance(s.ctx, s.address.Bytes(), s.bondDenom)
+			Expect(initialBalance.Amount).To(Equal(math.NewInt(4997609489499999900)))
+
+			withdrawCommissionArgs := defaultWithdrawCommissionArgs.
+				WithArgs(valAddr.String()).
+				WithGasPrice(gasPrice)
+
+			withdrawalCheck := passCheck.
+				WithExpEvents(distribution.EventTypeWithdrawValidatorCommission)
+
+			res, ethRes, err := contracts.CallContractAndCheckLogs(s.ctx, s.app, withdrawCommissionArgs, withdrawalCheck)
+			Expect(err).To(BeNil(), "error while calling the precompile")
+
+			var comm []cmn.Coin
+			err = s.precompile.UnpackIntoInterface(&comm, distribution.WithdrawValidatorCommissionMethod, ethRes.Ret)
+			Expect(err).To(BeNil())
+			Expect(len(comm)).To(Equal(1))
+			Expect(comm[0].Denom).To(Equal(s.bondDenom))
+			Expect(comm[0].Amount).To(Equal(expCommAmt))
+
+			finalBalance := s.app.BankKeeper.GetBalance(s.ctx, s.address.Bytes(), s.bondDenom)
+			fees := sdk.NewIntFromBigInt(gasPrice).MulRaw(res.GasUsed)
+			expFinal := initialBalance.Amount.Sub(fees)
+			Expect(finalBalance.Amount).To(Equal(expFinal), "expected final balance to be equal to the final balance after withdrawing commission")
+
+			// check that the commission was added to the withdrawer balance
+			finalWithdrawerBalance := s.app.BankKeeper.GetBalance(s.ctx, contractAddr.Bytes(), s.bondDenom)
+			Expect(finalWithdrawerBalance.Amount.BigInt()).To(Equal(expCommAmt))
 		})
 	})
 
@@ -769,6 +862,31 @@ var _ = Describe("Calling distribution precompile from another contract", func()
 			finalBalance := s.app.BankKeeper.GetBalance(s.ctx, contractAddr.Bytes(), s.bondDenom)
 			Expect(finalBalance.Amount.GT(initialBalance.Amount)).To(BeTrue(), "expected final balance to be greater than initial balance after withdrawing rewards")
 		})
+
+		It("should withdraw rewards successfully without origin check to a withdrawer address", func() {
+			withdrawerAddr, _ := testutiltx.NewAccAddressAndKey()
+
+			initialWithdrawerBalance := s.app.BankKeeper.GetBalance(s.ctx, withdrawerAddr.Bytes(), s.bondDenom)
+			Expect(initialWithdrawerBalance.Amount).To(Equal(sdk.ZeroInt()))
+
+			err := s.app.DistrKeeper.SetWithdrawAddr(s.ctx, contractAddr.Bytes(), withdrawerAddr.Bytes())
+			Expect(err).To(BeNil())
+
+			withdrawDelRewardsArgs := defaultWithdrawDelRewardsArgs.WithArgs(s.validators[0].OperatorAddress)
+
+			logCheckArgs := passCheck.WithExpEvents(distribution.EventTypeWithdrawDelegatorRewards)
+
+			_, _, err = contracts.CallContractAndCheckLogs(s.ctx, s.app, withdrawDelRewardsArgs, logCheckArgs)
+			Expect(err).To(BeNil(), "error while calling the smart contract: %v", err)
+
+			// withdrawer balance should increase with the rewards amt
+			finalWithdrawerBalance := s.app.BankKeeper.GetBalance(s.ctx, withdrawerAddr.Bytes(), s.bondDenom)
+			Expect(finalWithdrawerBalance.Amount.Equal(rewards)).To(BeTrue(), "expected final balance to be greater than initial balance after withdrawing rewards")
+
+			// delegator balance (contract) should remain unchanged
+			finalDelegatorBalance := s.app.BankKeeper.GetBalance(s.ctx, contractAddr.Bytes(), s.bondDenom)
+			Expect(finalDelegatorBalance.Amount.Equal(initialBalance.Amount)).To(BeTrue(), "expected delegator final balance remain unchanged after withdrawing rewards to withdrawer")
+		})
 	})
 
 	Context("withdrawValidatorCommission", func() {
@@ -851,6 +969,31 @@ var _ = Describe("Calling distribution precompile from another contract", func()
 			fees := gasPrice.Int64() * res.GasUsed
 			expFinal := initialBalance.Amount.Int64() + expValAmount - fees
 			Expect(finalBalance.Amount).To(Equal(math.NewInt(expFinal)), "expected final balance to be equal to initial balance + validator commission - fees")
+		})
+
+		It("should withdraw commission successfully to withdrawer address (contract)", func() {
+			initialWithdrawerBalance := s.app.BankKeeper.GetBalance(s.ctx, contractAddr.Bytes(), s.bondDenom)
+			Expect(initialWithdrawerBalance.Amount).To(Equal(sdk.ZeroInt()))
+
+			err := s.app.DistrKeeper.SetWithdrawAddr(s.ctx, s.address.Bytes(), contractAddr.Bytes())
+			Expect(err).To(BeNil())
+
+			withdrawValCommArgs := defaultWithdrawValCommArgs.
+				WithArgs(valAddr.String()).
+				WithGasPrice(gasPrice)
+			logCheckArgs := passCheck.
+				WithExpEvents(distribution.EventTypeWithdrawValidatorCommission)
+
+			res, _, err := contracts.CallContractAndCheckLogs(s.ctx, s.app, withdrawValCommArgs, logCheckArgs)
+			Expect(err).To(BeNil(), "error while calling the smart contract: %v", err)
+
+			finalWithdrawerBalance := s.app.BankKeeper.GetBalance(s.ctx, contractAddr.Bytes(), s.bondDenom)
+			Expect(finalWithdrawerBalance.Amount).To(Equal(math.NewInt(expValAmount)), "expected final balance to be equal to initial balance + validator commission")
+
+			finalBalance := s.app.BankKeeper.GetBalance(s.ctx, s.address.Bytes(), s.bondDenom)
+			fees := gasPrice.Int64() * res.GasUsed
+			expFinal := initialBalance.Amount.Int64() - fees
+			Expect(finalBalance.Amount).To(Equal(math.NewInt(expFinal)), "expected final balance to be equal to initial balance  - fees")
 		})
 	})
 
@@ -938,7 +1081,6 @@ var _ = Describe("Calling distribution precompile from another contract", func()
 			}...)
 
 			expectedBalance = sdk.Coin{Denom: utils.BaseDenom, Amount: math.NewInt(2e18)}
-
 			// populate default arguments
 			defaultClaimRewardsArgs = defaultCallArgs.WithMethodName(
 				"testClaimRewards",
@@ -959,9 +1101,13 @@ var _ = Describe("Calling distribution precompile from another contract", func()
 		})
 
 		It("should withdraw rewards successfully to a different address without origin check", func() {
-			initialBalance := s.app.BankKeeper.GetBalance(s.ctx, s.address.Bytes(), s.bondDenom)
+			withdrawerAddr, _ := testutiltx.NewAccAddressAndKey()
+			initialWithdrawerBalance := s.app.BankKeeper.GetBalance(s.ctx, withdrawerAddr.Bytes(), s.bondDenom)
+			Expect(initialWithdrawerBalance.Amount).To(Equal(sdk.ZeroInt()))
 
-			err := s.app.DistrKeeper.SetWithdrawAddr(s.ctx, contractAddr.Bytes(), s.address.Bytes())
+			initialDelegatorBalance := s.app.BankKeeper.GetBalance(s.ctx, contractAddr.Bytes(), s.bondDenom)
+
+			err := s.app.DistrKeeper.SetWithdrawAddr(s.ctx, contractAddr.Bytes(), withdrawerAddr.Bytes())
 			Expect(err).To(BeNil())
 
 			claimRewardsArgs := defaultClaimRewardsArgs.WithArgs(contractAddr, uint32(2))
@@ -971,9 +1117,13 @@ var _ = Describe("Calling distribution precompile from another contract", func()
 			_, _, err = contracts.CallContractAndCheckLogs(s.ctx, s.app, claimRewardsArgs, logCheckArgs)
 			Expect(err).To(BeNil(), "error while calling the smart contract: %v", err)
 
-			// balance should increase
-			finalBalance := s.app.BankKeeper.GetBalance(s.ctx, s.address.Bytes(), s.bondDenom)
-			Expect(finalBalance.Amount.GT(initialBalance.Amount)).To(BeTrue(), "expected final balance to be greater than initial balance after withdrawing rewards")
+			// withdrawer balance should increase
+			finalWithdrawerBalance := s.app.BankKeeper.GetBalance(s.ctx, withdrawerAddr.Bytes(), s.bondDenom)
+			Expect(finalWithdrawerBalance.Equal(expectedBalance)).To(BeTrue(), "expected final withdrawer balance to be greater than initial balance after withdrawing rewards")
+
+			// delegator (contract) balance should remain unchanged
+			finalDelegatorBalance := s.app.BankKeeper.GetBalance(s.ctx, contractAddr.Bytes(), s.bondDenom)
+			Expect(finalDelegatorBalance.Amount.Equal(initialDelegatorBalance.Amount)).To(BeTrue(), "expected final delegator balance to same as initial balance after withdrawing rewards")
 		})
 	})