+++ CHANGELOG.md
@@ -55,6 +55,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (staking) [#2051](https://github.com/evmos/evmos/pull/2051) Implement the `EditValidator` function for staking precompiled contract.
 - (evm) [#2538](https://github.com/evmos/evmos/pull/2538) Add Permissions Policy for permissioned EVM.
 - (erc20) [#2609] (https://github.com/evmos/evmos/pull/2609) Remove STRv2 tracking logic.
+- (distribution-precompile) [#2614] (https://github.com/evmos/evmos/pull/2614) Add withdrawer address check in transactions.
 
 ### Bug Fixes
 
@@ -106,6 +107,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 - (app) [#2373](https://github.com/evmos/evmos/pull/2373) Remove ChainID constraint.
 - (rpc) [#2375](https://github.com/evmos/evmos/pull/2375) Add support for allow-insecure-unlock.
+- (tests) [#2575](https://github.com/evmos/evmos/pull/2575) Add integration test for precompile tx `CreateValidator`.
 - (evm) [#2581](https://github.com/evmos/evmos/pull/2581) Fix wrong error message returned for DynamicFee txs.
 
 ### State Machine Breaking